pyproject.toml
[project]
name = "pypgwire"
version = "0.1.0"
description = "Postgres wire protocol in Python"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "asyncpg>=0.31.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pyclean",
    "psycopg2-binary>=2.9.11",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = ["--import-mode=importlib",]

